requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI Integration
google-generativeai>=0.8.0

# Middleware & Utilities
python-multipart==0.0.6
slowapi==0.1.9

# Testing (optional)
requests>=2.31.0
